Back/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Conexiones a bases de datos
google-api-core
google-cloud-bigquery==3.13.0
pymongo==4.6.1
neo4j==5.15.0

# Utilidades
python-multipart==0.0.6
python-dotenv==1.0.0

# Logging y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

pytest==7.4.3
httpx==0.25.2